Commit stray diff

+++ src/net/psj/Walls/WallBasic.java
@@ -5,6 +5,6 @@ import net.psj.Interface.MenuData;
 public class WallBasic extends Wall {
 
 	public WallBasic() {
-		super("Wall", new float[] { 0.6f, 0.6f, 0.6f }, 1, MenuData.MENU_WALLS);
+		super("Wall", new float[] { 0.6f, 0.6f, 0.6f }, (short) 1, MenuData.MENU_WALLS);
 	}
 }
